Nav and Dashboard container are now their own components and reused by each page
