+++ client/src/components/Nav.js
@@ -1,0 +1,205 @@
+import { Fragment, useState, useContext } from 'react'
+import { Dialog, Transition } from '@headlessui/react'
+import {
+  CalendarIcon,
+  ChartBarIcon,
+  FolderIcon,
+  HomeIcon,
+  InboxIcon,
+  MenuIcon,
+  UsersIcon,
+  XIcon,
+  BookOpenIcon
+} from '@heroicons/react/outline'
+import { AuthContext } from './context/auth'
+
+
+
+const navigation = [
+  { name: 'Dashboard', href: '/', icon: HomeIcon, current: true },
+  { name: 'All books', href: '/all-books', icon: BookOpenIcon, current: false },
+  { name: 'View profile', href: '/profile', icon: UsersIcon, current: false },
+  // { name: 'Calendar', href: '#', icon: CalendarIcon, current: false },
+  // { name: 'Documents', href: '#', icon: InboxIcon, current: false },
+  // { name: 'Reports', href: '#', icon: ChartBarIcon, current: false },
+]
+
+function classNames(...classes) {
+  return classes.filter(Boolean).join(' ')
+}
+
+export const Nav = () => {
+  const { user, logout } = useContext(AuthContext)
+  const [sidebarOpen, setSidebarOpen] = useState(false)
+  return (
+    <>
+      <Transition.Root show={sidebarOpen} as={Fragment}>
+      <Dialog
+        as="div"
+        static
+        className="fixed inset-0 flex z-40 md:hidden"
+        open={sidebarOpen}
+        onClose={setSidebarOpen}
+      >
+        <Transition.Child
+          as={Fragment}
+          enter="transition-opacity ease-linear duration-300"
+          enterFrom="opacity-0"
+          enterTo="opacity-100"
+          leave="transition-opacity ease-linear duration-300"
+          leaveFrom="opacity-100"
+          leaveTo="opacity-0"
+        >
+          <Dialog.Overlay className="fixed inset-0 bg-gray-600 bg-opacity-75" />
+        </Transition.Child>
+        <Transition.Child
+          as={Fragment}
+          enter="transition ease-in-out duration-300 transform"
+          enterFrom="-translate-x-full"
+          enterTo="translate-x-0"
+          leave="transition ease-in-out duration-300 transform"
+          leaveFrom="translate-x-0"
+          leaveTo="-translate-x-full"
+        >
+          <div className="relative flex-1 flex flex-col max-w-xs w-full bg-gray-800">
+            <Transition.Child
+              as={Fragment}
+              enter="ease-in-out duration-300"
+              enterFrom="opacity-0"
+              enterTo="opacity-100"
+              leave="ease-in-out duration-300"
+              leaveFrom="opacity-100"
+              leaveTo="opacity-0"
+            >
+              <div className="absolute top-0 right-0 -mr-12 pt-2">
+                <button
+                  className="ml-1 flex items-center justify-center h-10 w-10 rounded-full focus:outline-none focus:ring-2 focus:ring-inset focus:ring-white"
+                  onClick={() => setSidebarOpen(false)}
+                >
+                  <span className="sr-only">Close sidebar</span>
+                  <XIcon className="h-6 w-6 text-white" aria-hidden="true" />
+                </button>
+              </div>
+            </Transition.Child>
+            <div className="flex-1 h-0 pt-5 pb-4 overflow-y-auto">
+              <div className="flex-shrink-0 flex items-center px-4">
+                <img
+                  className="h-8 w-auto"
+                  src="https://tailwindui.com/img/logos/workflow-logo-indigo-500-mark-white-text.svg"
+                  alt="Workflow"
+                />
+              </div>
+              <nav className="mt-5 px-2 space-y-1">
+                {navigation.map((item) => (
+                  <a
+                    key={item.name}
+                    href={item.href}
+                    className={classNames(
+                      item.current ? 'bg-gray-900 text-white' : 'text-gray-300 hover:bg-gray-700 hover:text-white',
+                      'group flex items-center px-2 py-2 text-base font-medium rounded-md'
+                    )}
+                  >
+                    <item.icon
+                      className={classNames(
+                        item.current ? 'text-gray-300' : 'text-gray-400 group-hover:text-gray-300',
+                        'mr-4 h-6 w-6'
+                      )}
+                      aria-hidden="true"
+                    />
+                    {item.name}
+                  </a>
+                ))}
+              </nav>
+            </div>
+            <div className="flex-shrink-0 flex bg-gray-700 p-4">
+              <a href="#" className="flex-shrink-0 group block">
+                <div className="flex items-center">
+                  <div>
+                    <img
+                      className="inline-block h-10 w-10 rounded-full"
+                      src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
+                      alt=""
+                    />
+                  </div>
+                  <div className="ml-3">
+                    <p className="text-base font-medium text-white">Tom Cook</p>
+                    <p className="text-sm font-medium text-gray-400 group-hover:text-gray-300">View profile</p>
+                  </div>
+                </div>
+              </a>
+            </div>
+          </div>
+        </Transition.Child>
+        <div className="flex-shrink-0 w-14">{/* Force sidebar to shrink to fit close icon */}</div>
+      </Dialog>
+    </Transition.Root>
+
+    {/* Static sidebar for desktop */}
+    <div className="hidden md:flex md:flex-shrink-0">
+      <div className="flex flex-col w-64">
+        <div className="flex flex-col h-0 flex-1 bg-gray-800">
+          <div className="flex-1 flex flex-col pt-5 pb-4 overflow-y-auto">
+            <div className="flex items-center flex-shrink-0 px-4">
+              <img
+                className="h-8 w-auto"
+                src="https://tailwindui.com/img/logos/workflow-logo-indigo-500-mark-white-text.svg"
+                alt="Workflow"
+              />
+            </div>
+            <nav className="mt-5 flex-1 px-2 bg-gray-800 space-y-1">
+              {navigation.map((item) => (
+                <a
+                  key={item.name}
+                  href={item.href}
+                  className={classNames(
+                    item.current ? 'bg-gray-900 text-white' : 'text-gray-300 hover:bg-gray-700 hover:text-white',
+                    'group flex items-center px-2 py-2 text-sm font-medium rounded-md'
+                  )}
+                >
+                  <item.icon
+                    className={classNames(
+                      item.current ? 'text-gray-300' : 'text-gray-400 group-hover:text-gray-300',
+                      'mr-3 h-6 w-6'
+                    )}
+                    aria-hidden="true"
+                  />
+                  {item.name}
+                </a>
+              ))}
+            </nav>
+          </div>
+          <div className="flex-shrink-0 flex bg-gray-700 p-4">
+            <a href="#" className="flex-shrink-0 w-full group block">
+              <div className="flex items-center">
+                <div>
+                  <img
+                    className="inline-block h-9 w-9 rounded-full"
+                    src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
+                    alt=""
+                  />
+                </div>
+                <div className="ml-3">
+                  <p className="text-sm font-medium text-white">{user.firstName} {user.lastName}</p>
+                  <p className="text-xs font-medium text-gray-300 group-hover:text-gray-200">View profile</p>
+                  <button onClick={logout} className="text-xs font-medium text-gray-300 group-hover:text-gray-200">Log out</button>
+                </div>
+              </div>
+            </a>
+          </div>
+        </div>
+      </div>
+    </div>
+    {/* <div className="flex flex-col w-0 flex-1 overflow-hidden"> */}
+      <div className="md:hidden pl-1 pt-1 sm:pl-3 sm:pt-3">
+        <button
+          className="-ml-0.5 -mt-0.5 h-12 w-12 inline-flex items-center justify-center rounded-md text-gray-500 hover:text-gray-900 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500"
+          onClick={() => setSidebarOpen(true)}
+        >
+          <span className="sr-only">Open sidebar</span>
+          <MenuIcon className="h-6 w-6" aria-hidden="true" />
+        </button>
+      </div>
+    {/* </div> */}
+   </> 
+  )
+}
