+++ client/src/components/pages/AllBooks.js
@@ -1,15 +1,16 @@
 import { useQuery } from "@apollo/client"
 import gql from "graphql-tag"
 import BookTable from '../BookTable'
+import { Page } from "./Page";
 
 const GET_ALL_BOOKS = gql`
   query getAllBooks {
     getAllBooks {
-    title
-    description
-    author
-    price
-  }
+      title
+      description
+      author
+      price
+    }
   }
 `;
 
@@ -18,10 +19,10 @@ export const AllBooks = () => {
 
   if(loading) return <p>Loading...</p>
   if(error) return <p>Error: {error.message}</p>
-    console.log(data)
+  
   return (
-    <div>
-      <BookTable bookData={data.getAllBooks}/>
-    </div>
+    <Page pageTitle={'All Books'}>
+      <BookTable bookData={data.getAllBooks} />
+    </Page>
   )
 }
